added a new component for /haulingships that shows a list of all haulers with an edit button next to each listing
